fix(acbs/utils.py): use ~dirty suffix for uncommited tree
Previously, when we used -dirty, the second dash (-) was considered the actual
delimiter for REL. If the previous package version already contained a REL
(now the first dash), then the segment between the beginning and the second
dash would be considered the VER - making the package version higher than
any comparable candidate without the -dirty suffix.

Revise -dirty as ~dirty to close this loophole.